+++ trunk/projects/Proteins/Rosetta/Sampler/GaussianMultivariateSampler.h
@@ -18,67 +18,67 @@ namespace lbcpp
 class GaussianMultivariateSampler;
 typedef ReferenceCountedObjectPtr<GaussianMultivariateSampler> GaussianMultivariateSamplerPtr;
 
-class GaussianMultivariateSampler: public ContinuousSampler
+class GaussianMultivariateSampler : public ContinuousSampler
 {
 public:
-  GaussianMultivariateSampler() :
-    ContinuousSampler(0, 0), maxIt(100), tolerance(0.01), numVariables(0), numClusters(0)
+  GaussianMultivariateSampler()
+    : ContinuousSampler(0, 0), maxIt(100), tolerance(0.01), numVariables(0), numClusters(0)
   {
   }
 
-  GaussianMultivariateSampler(int maxIt, double tolerance, MatrixPtr& initialProbabilities,
-      std::vector<MatrixPtr>& initialMean, std::vector<MatrixPtr>& initialStd) :
-    ContinuousSampler(0.0, 1.0), maxIt(maxIt), tolerance(tolerance), numVariables(
+  GaussianMultivariateSampler(size_t maxIt, double tolerance, MatrixPtr& initialProbabilities,
+      std::vector<MatrixPtr>& initialMean, std::vector<MatrixPtr>& initialStd)
+    : ContinuousSampler(0.0, 1.0), maxIt(maxIt), tolerance(tolerance), numVariables(
         initialMean[0]->getNumRows()), numClusters(initialProbabilities->getNumRows())
   {
     probabilities = new DoubleMatrix(numClusters, 1, 1.0 / numClusters);
     double norm = 0;
-    for (int j = 0; j < numClusters; j++)
+    for (size_t j = 0; j < numClusters; j++)
     {
       probabilities->setElement(j, 0, initialProbabilities->getElement(j, 0));
-      norm += probabilities->getElement(j,0).getDouble();
+      norm += probabilities->getElement(j, 0).getDouble();
     }
     // ensures that variables are normalized
     double factor = 1.0 / norm;
-    for (int j = 0; j < numClusters; j++)
-      probabilities->setElement(j, 0, probabilities->getElement(j, 0).getDouble() * factor );
+    for (size_t j = 0; j < numClusters; j++)
+      probabilities->setElement(j, 0, probabilities->getElement(j, 0).getDouble() * factor);
 
     means = std::vector<MatrixPtr>(numClusters);
     covarianceMatrix = std::vector<MatrixPtr>(numClusters);
-    for (int j = 0; j < numClusters; j++)
+    for (size_t j = 0; j < numClusters; j++)
     {
       means[j] = new DoubleMatrix(numVariables, 1, 0.0);
-      for (int k = 0; k < numVariables; k++)
+      for (size_t k = 0; k < numVariables; k++)
         means[j]->setElement(k, 0, initialMean[j]->getElement(k, 0));
 
       covarianceMatrix[j] = new DoubleMatrix(numVariables, numVariables, 0.0);
-      for (int k = 0; k < covarianceMatrix[j]->getNumRows(); k++)
-        for (int l = 0; l < covarianceMatrix[j]->getNumColumns(); l++)
-            covarianceMatrix[j]->setElement(k, l, initialStd[j]->getElement(k, l));
+      for (size_t k = 0; k < covarianceMatrix[j]->getNumRows(); k++)
+        for (size_t l = 0; l < covarianceMatrix[j]->getNumColumns(); l++)
+          covarianceMatrix[j]->setElement(k, l, initialStd[j]->getElement(k, l));
     }
   }
 
-  GaussianMultivariateSampler(const GaussianMultivariateSampler& copy) :
-    ContinuousSampler(copy.mean, copy.std), maxIt(copy.maxIt), tolerance(copy.tolerance),
+  GaussianMultivariateSampler(const GaussianMultivariateSampler& copy)
+    : ContinuousSampler(copy.mean, copy.std), maxIt(copy.maxIt), tolerance(copy.tolerance),
         numVariables(copy.numVariables), numClusters(copy.numClusters)
   {
     probabilities = new DoubleMatrix(copy.probabilities->getNumRows(), 1);
-    for (int j = 0; j < copy.probabilities->getNumRows(); j++)
+    for (size_t j = 0; j < copy.probabilities->getNumRows(); j++)
       probabilities->setElement(j, 0, copy.probabilities->getElement(j, 0));
 
     means = std::vector<MatrixPtr>(copy.numClusters);
     covarianceMatrix = std::vector<MatrixPtr>(copy.numClusters);
 
-    for (int j = 0; j < copy.numClusters; j++)
+    for (size_t j = 0; j < copy.numClusters; j++)
     {
       means[j] = new DoubleMatrix(copy.means[j]->getNumRows(), 1);
-      for (int k = 0; k < copy.means[j]->getNumRows(); k++)
+      for (size_t k = 0; k < copy.means[j]->getNumRows(); k++)
         means[j]->setElement(k, 0, copy.means[j]->getElement(k, 0));
 
       covarianceMatrix[j] = new DoubleMatrix(copy.covarianceMatrix[0]->getNumRows(),
           copy.covarianceMatrix[0]->getNumColumns(), 0);
-      for (int k = 0; k < copy.covarianceMatrix[j]->getNumRows(); k++)
-        for (int l = 0; l < copy.covarianceMatrix[j]->getNumColumns(); l++)
+      for (size_t k = 0; k < copy.covarianceMatrix[j]->getNumRows(); k++)
+        for (size_t l = 0; l < copy.covarianceMatrix[j]->getNumColumns(); l++)
           covarianceMatrix[j]->setElement(k, l, copy.covarianceMatrix[j]->getElement(k, l));
     }
   }
@@ -97,18 +97,18 @@ public:
   {
     ClassPtr actionClass = denseDoubleVectorClass(positiveIntegerEnumerationEnumeration);
     DenseDoubleVectorPtr proba = new DenseDoubleVector(actionClass, numClusters, 0.0);
-    for (int i = 0; i < numClusters; i++)
+    for (size_t i = 0; i < numClusters; i++)
       proba->setValue(i, probabilities->getElement(i, 0).getDouble());
-    int distrib = random->sampleWithProbabilities(proba->getValues());
+    size_t distrib = random->sampleWithProbabilities(proba->getValues());
 
     MatrixPtr normalSeed = new DoubleMatrix(numVariables, 1, 0.0);
-    for (int i = 0; i < numVariables; i++)
+    for (size_t i = 0; i < numVariables; i++)
       normalSeed->setElement(i, 0, Variable(random->sampleDoubleFromGaussian(0, 1)));
 
     MatrixPtr chol = choleskyDecomposition(covarianceMatrix[distrib]);
     MatrixPtr result = matrixProduct(chol, normalSeed);
 
-    for (int i = 0; i < result->getNumRows(); i++)
+    for (size_t i = 0; i < result->getNumRows(); i++)
       result->setElement(i, 0, Variable(result->getElement(i, 0).getDouble()
           + means[distrib]->getElement(i, 0).getDouble()));
 
@@ -125,13 +125,13 @@ public:
     }
 
     std::vector<MatrixPtr> vals(dataset.size());
-    for (int i = 0; i < dataset.size(); i++)
+    for (size_t i = 0; i < dataset.size(); i++)
       vals[i] = dataset[i].first.getObjectAndCast<Matrix> ();
 
     double oldLL = computeLogLikelihood(vals, probabilities, means, covarianceMatrix);
     double newLL = oldLL + 2 * oldLL * tolerance;
 
-    for (int i = 0; ((std::abs((oldLL - newLL) / oldLL) > tolerance) && (i < maxIt)); i++)
+    for (size_t i = 0; ((std::abs((oldLL - newLL) / oldLL) > tolerance) && (i < maxIt)); i++)
     {
       MatrixPtr gammas;
       MatrixPtr newProbabilities;
@@ -154,17 +154,17 @@ public:
     // probabilities
     probabilitiesToUpdate = new DoubleMatrix(newProbabilities->getNumRows(),
         newProbabilities->getNumColumns());
-    for (int j = 0; j < newProbabilities->getNumRows(); j++)
+    for (size_t j = 0; j < newProbabilities->getNumRows(); j++)
       probabilitiesToUpdate->setElement(j, 0, Variable(
           newProbabilities->getElement(j, 0).getDouble() / (double)dataset.size()));
 
     // means
     meansToUpdate = std::vector<MatrixPtr>(newProbabilities->getNumRows());
-    for (int j = 0; j < meansToUpdate.size(); j++)
+    for (size_t j = 0; j < meansToUpdate.size(); j++)
     {
       MatrixPtr tempSum = new DoubleMatrix(dataset[0]->getNumRows(), 1);
 
-      for (int i = 0; i < dataset.size(); i++)
+      for (size_t i = 0; i < dataset.size(); i++)
         tempSum = addMatrix(tempSum, scalarMultiplyMatrix(dataset[i],
             gammas->getElement(i, j).getDouble()));
       meansToUpdate[j] = scalarMultiplyMatrix(tempSum, 1.0
@@ -173,11 +173,11 @@ public:
 
     // covariances
     covariancesToUpdate = std::vector<MatrixPtr>(newProbabilities->getNumRows());
-    for (int j = 0; j < covariancesToUpdate.size(); j++)
+    for (size_t j = 0; j < covariancesToUpdate.size(); j++)
     {
       MatrixPtr tempSub = new DoubleMatrix(dataset[0]->getNumRows(), 1);
       MatrixPtr acc = new DoubleMatrix(dataset[0]->getNumRows(), dataset[0]->getNumRows());
-      for (int i = 0; i < dataset.size(); i++)
+      for (size_t i = 0; i < dataset.size(); i++)
       {
         tempSub = subtractMatrix(dataset[i], meansToUpdate[j]);
         acc = addMatrix(acc, scalarMultiplyMatrix(matrixProduct(tempSub, transposeMatrix(tempSub)),
@@ -197,13 +197,13 @@ public:
   {
     gammas = new DoubleMatrix(dataset.size(), means.size());
     std::vector<MatrixPtr> invCov(covariances.size());
-    for (int i = 0; i < covariances.size(); i++)
+    for (size_t i = 0; i < covariances.size(); i++)
       invCov[i] = inverseMatrix(covariances[i]);
 
-    for (int i = 0; i < gammas->getNumRows(); i++)
+    for (size_t i = 0; i < gammas->getNumRows(); i++)
     {
       double norm = 0;
-      for (int j = 0; j < gammas->getNumColumns(); j++)
+      for (size_t j = 0; j < gammas->getNumColumns(); j++)
       {
         double increment = probabilities->getElement(j, 0).getDouble() * computeProbability(
             dataset[i], means[j], covariances[j], invCov[j]);
@@ -212,15 +212,15 @@ public:
       }
 
       double invNorm = 1.0 / norm;
-      for (int j = 0; j < gammas->getNumColumns(); j++)
+      for (size_t j = 0; j < gammas->getNumColumns(); j++)
         gammas->setElement(i, j, Variable(gammas->getElement(i, j).getDouble() * invNorm));
     }
 
     newProbabilities = new DoubleMatrix(probabilities->getNumRows(), 1);
-    for (int j = 0; j < gammas->getNumColumns(); j++)
+    for (size_t j = 0; j < gammas->getNumColumns(); j++)
     {
       double tempSum = 0;
-      for (int i = 0; i < gammas->getNumRows(); i++)
+      for (size_t i = 0; i < gammas->getNumRows(); i++)
         tempSum += gammas->getElement(i, j).getDouble();
       newProbabilities->setElement(j, 0, Variable(tempSum));
     }
@@ -234,15 +234,15 @@ public:
       std::vector<MatrixPtr>& means, std::vector<MatrixPtr>& covariances)
   {
     std::vector<MatrixPtr> invCov(covariances.size());
-    for (int i = 0; i < covariances.size(); i++)
+    for (size_t i = 0; i < covariances.size(); i++)
       invCov[i] = inverseMatrix(covariances[i]);
 
     double ll = 0;
 
-    for (int i = 0; i < dataset.size(); i++)
+    for (size_t i = 0; i < dataset.size(); i++)
     {
       double tempSum = 0;
-      for (int j = 0; j < probabilities->getNumRows(); j++)
+      for (size_t j = 0; j < probabilities->getNumRows(); j++)
         tempSum += probabilities->getElement(j, 0).getDouble() * computeProbability(dataset[i],
             means[j], covariances[j], invCov[j]);
 
@@ -276,7 +276,7 @@ public:
     return numerator * denominator;
   }
 
-  double computeNormalizationFactor(int dimension, MatrixPtr& covariance)
+  double computeNormalizationFactor(size_t dimension, MatrixPtr& covariance)
   {
     return 1.0 / (std::pow(2 * M_PI, (double)dimension / 2.0) * std::pow(determinantMatrix(
         covariance), 0.5));
@@ -292,29 +292,30 @@ public:
   MatrixPtr choleskyDecomposition(const MatrixPtr& a) const
   {
     MatrixPtr temp = new DoubleMatrix(a->getNumRows(), a->getNumColumns(), 0.0);
-    for (int i = 0; i < a->getNumRows(); i++)
-      for (int j = 0; j < a->getNumColumns(); j++)
+    for (size_t i = 0; i < a->getNumRows(); i++)
+      for (size_t j = 0; j < a->getNumColumns(); j++)
         temp->setElement(i, j, Variable(a->getElement(i, j).getDouble()));
 
     // Decomposition
     temp->setElement(0, 0, Variable(std::sqrt(std::abs(temp->getElement(0, 0).getDouble()))));
-    int N = temp->getNumColumns();
-    for (int i = 1; i < N; i++)
+    size_t N = temp->getNumColumns();
+    for (size_t i = 1; i < N; i++)
       temp->setElement(i, 0, Variable(temp->getElement(i, 0).getDouble()
           / temp->getElement(0, 0).getDouble()));
 
-    for (int k = 1; k < (N - 1); k++)
+    for (size_t k = 1; k < (N - 1); k++)
     {
       double sum1 = 0;
 
-      for (int j = 0; j < k; j++)
+      for (size_t j = 0; j < k; j++)
         sum1 += std::pow(temp->getElement(k, j).getDouble(), 2.0);
-      temp->setElement(k, k, Variable(std::sqrt(std::abs(temp->getElement(k, k).getDouble() - sum1))));
+      temp->setElement(k, k, Variable(
+          std::sqrt(std::abs(temp->getElement(k, k).getDouble() - sum1))));
 
-      for (int i = k + 1; i < N; i++)
+      for (size_t i = k + 1; i < N; i++)
       {
         double sum3 = 0;
-        for (int j = 0; j < k; j++)
+        for (size_t j = 0; j < k; j++)
           sum3 += temp->getElement(i, j).getDouble() * temp->getElement(k, j).getDouble();
         temp->setElement(i, k, Variable((1.0 / temp->getElement(k, k).getDouble())
             * (temp->getElement(i, k).getDouble() - sum3)));
@@ -322,14 +323,14 @@ public:
     }
 
     double sum2 = 0;
-    for (int j = 0; j < (N - 1); j++)
+    for (size_t j = 0; j < (N - 1); j++)
       sum2 += std::pow(temp->getElement(N - 1, j).getDouble(), 2.0);
-    temp->setElement(N - 1, N - 1, Variable(std::sqrt(std::abs(temp->getElement(N - 1, N - 1).getDouble()
-        - sum2))));
+    temp->setElement(N - 1, N - 1, Variable(std::sqrt(std::abs(
+        temp->getElement(N - 1, N - 1).getDouble() - sum2))));
 
     // Set the upper part to 0
-    for (int i = 0; i < a->getNumRows(); i++)
-      for (int j = i + 1; j < a->getNumColumns(); j++)
+    for (size_t i = 0; i < a->getNumRows(); i++)
+      for (size_t j = i + 1; j < a->getNumColumns(); j++)
         temp->setElement(i, j, Variable(0.0));
 
     return temp;
@@ -339,9 +340,9 @@ public:
   {
     MatrixPtr temp = new DoubleMatrix(a->getNumRows(), b->getNumColumns(), 0.0);
 
-    for (int i = 0; i < a->getNumRows(); i++)
-      for (int j = 0; j < b->getNumColumns(); j++)
-        for (int k = 0; k < a->getNumColumns(); k++)
+    for (size_t i = 0; i < a->getNumRows(); i++)
+      for (size_t j = 0; j < b->getNumColumns(); j++)
+        for (size_t k = 0; k < a->getNumColumns(); k++)
         {
           double value = temp->getElement(i, j).getDouble();
           value += a->getElement(i, k).getDouble() * b->getElement(k, j).getDouble();
@@ -353,9 +354,9 @@ public:
 
   void printMatrix(MatrixPtr m) const
   {
-    for (int i = 0; i < m->getNumRows(); i++)
+    for (size_t i = 0; i < m->getNumRows(); i++)
     {
-      for (int j = 0; j < m->getNumColumns(); j++)
+      for (size_t j = 0; j < m->getNumColumns(); j++)
         std::cout << m->getElement(i, j).getDouble() << " : ";
       std::cout << std::endl;
     }
@@ -365,8 +366,8 @@ public:
   MatrixPtr transposeMatrix(MatrixPtr a) const
   {
     MatrixPtr temp = new DoubleMatrix(a->getNumColumns(), a->getNumRows(), 0.0);
-    for (int i = 0; i < a->getNumRows(); i++)
-      for (int j = 0; j < a->getNumColumns(); j++)
+    for (size_t i = 0; i < a->getNumRows(); i++)
+      for (size_t j = 0; j < a->getNumColumns(); j++)
         temp->setElement(j, i, a->getElement(i, j));
 
     return temp;
@@ -375,8 +376,8 @@ public:
   MatrixPtr addMatrix(MatrixPtr a, MatrixPtr b)
   {
     MatrixPtr temp = new DoubleMatrix(a->getNumRows(), a->getNumColumns(), 0.0);
-    for (int i = 0; i < a->getNumRows(); i++)
-      for (int j = 0; j < a->getNumColumns(); j++)
+    for (size_t i = 0; i < a->getNumRows(); i++)
+      for (size_t j = 0; j < a->getNumColumns(); j++)
         temp->setElement(i, j, Variable(a->getElement(i, j).getDouble()
             + b->getElement(i, j).getDouble()));
 
@@ -386,8 +387,8 @@ public:
   MatrixPtr subtractMatrix(MatrixPtr a, MatrixPtr b)
   {
     MatrixPtr temp = new DoubleMatrix(a->getNumRows(), a->getNumColumns(), 0.0);
-    for (int i = 0; i < a->getNumRows(); i++)
-      for (int j = 0; j < a->getNumColumns(); j++)
+    for (size_t i = 0; i < a->getNumRows(); i++)
+      for (size_t j = 0; j < a->getNumColumns(); j++)
         temp->setElement(i, j, Variable(a->getElement(i, j).getDouble()
             - b->getElement(i, j).getDouble()));
 
@@ -397,8 +398,8 @@ public:
   MatrixPtr scalarMultiplyMatrix(MatrixPtr a, double b)
   {
     MatrixPtr temp = new DoubleMatrix(a->getNumRows(), a->getNumColumns(), 0.0);
-    for (int i = 0; i < a->getNumRows(); i++)
-      for (int j = 0; j < a->getNumColumns(); j++)
+    for (size_t i = 0; i < a->getNumRows(); i++)
+      for (size_t j = 0; j < a->getNumColumns(); j++)
         temp->setElement(i, j, Variable(b * a->getElement(i, j).getDouble()));
 
     return temp;
@@ -424,10 +425,10 @@ public:
   double normOneMatrix(MatrixPtr a)
   {
     double max = -1.0;
-    for (int j = 0; j < a->getNumColumns(); j++)
+    for (size_t j = 0; j < a->getNumColumns(); j++)
     {
       double sum = 0;
-      for (int i = 0; i < a->getNumRows(); i++)
+      for (size_t i = 0; i < a->getNumRows(); i++)
       {
         sum += std::abs(a->getElement(i, j).getDouble());
       }
@@ -450,9 +451,9 @@ public:
 protected:
   friend class GaussianMultivariateSamplerClass;
   double tolerance;
-  int maxIt;
-  int numVariables;
-  int numClusters;
+  size_t maxIt;
+  size_t numVariables;
+  size_t numClusters;
   MatrixPtr probabilities;
   std::vector<MatrixPtr> means;
   std::vector<MatrixPtr> covarianceMatrix;
